+++ plugins/woocommerce-admin/client/analytics/settings/setting.js
@@ -23,7 +23,7 @@ class Setting extends Component {
 	}
 
 	renderInput = () => {
-		const { handleChange, name, inputText, inputType, options, value } = this.props;
+		const { handleChange, name, inputText, inputType, options, value, component } = this.props;
 		const { disabled } = this.state;
 		const id = uniqueId( name );
 
@@ -54,6 +54,9 @@ class Setting extends Component {
 						{ inputText }
 					</Button>
 				);
+			case 'component':
+				const SettingComponent = component;
+				return <SettingComponent value={ value } onChange={ handleChange } { ...this.props } />;
 			case 'text':
 			default:
 				return (
@@ -150,7 +153,7 @@ Setting.propTypes = {
 	/**
 	 * Type of input to use; defaults to a text input.
 	 */
-	inputType: PropTypes.oneOf( [ 'button', 'checkbox', 'checkboxGroup', 'text' ] ),
+	inputType: PropTypes.oneOf( [ 'button', 'checkbox', 'checkboxGroup', 'text', 'component' ] ),
 	/**
 	 * Label used for describing the setting.
 	 */